move is_disconnected_patch to a new file - vllm/entrypoints/utils.py
